Time: 17 ms (7.81%), Space: 40 MB (83.45%) - LeetHub

+++ 441-arranging-coins/441-arranging-coins.java
@@ -1,6 +1,6 @@
 class Solution {
     public int arrangeCoins(int n) {
-        double answer = 0;
+        long answer = 0;
         for (int i=1; i<=n; i++) {
             answer += i;
             if (answer > n) {
